+++ mods/src/multiplayer/inventory.cpp
@@ -64,6 +64,7 @@ static void send_inventory(LocalPlayer *self) {
 
     // Send
     self->minecraft->rak_net_instance->send(*(Packet *) packet);
+    packet->destructor_deleting();
 }
 static void LocalPlayer_tick_injection(LocalPlayer_tick_t original, LocalPlayer *self) {
     // Call Original Method